added comments and docstrings and moved command to commands rather than logic
